Fix failing tests due to changes in MVN package
MVN >= 5.9 now uses Anderson-Darling for univariate tests, which
requires a sample of at least 8 elements per group. Previously
it was using Shapiro-Wilk, which does not have this "limitation".
Since micompr ignores the results of the univariate tests, we
explicitly invoke MVN with univariateTest = "SW" so that the
tests continue testing what they were testing before this change
in MVN.